+++ private/system/classes/menu.class.php
@@ -416,6 +416,7 @@ class menuElement {
             return NULL;
         }
         switch ( $this->type ) {
+
             case ET_SUB_MENU :
                 $this->replace_macros();
                 break;
@@ -476,7 +477,16 @@ class menuElement {
                 $this->url = '';
                 switch ($this->subtype) {
                     case USER_MENU :
+// if anonymous user - show login entry
+                        if ( COM_isAnonUser() ) {
+                            $this->label = $LANG01[58];
+                            $this->url   = $_CONF['site_url'] . '/users.php';
+                            $this->target = '';
+                            break;
+                        }
+// logged-in user see My Account entry
                         $item_array = getUserMenu();
+                        $this->label = $LANG01[47];
                         break;
 
                     case ADMIN_MENU :